refactored audio type, removed useless component
