+++ src/components/types/audio.tsx
@@ -1,0 +1,10 @@
+interface Audio {
+    title: string
+    artist: string
+    album: string
+    id: number
+    duration: number
+    cover: number[] // byte array
+}
+
+export type { Audio }
