+++ src/pages/index.tsx
@@ -1,12 +1,12 @@
 import type { NextPage } from "next"
 import Head from "next/head"
 import { useState } from "react"
-import Link from "next/link"
 import { playlists } from "@/components/data/playlists"
 import { MenubarDemo } from "@/components/MenubarDemo"
 import { Player } from "@/components/Player"
-import { Audio, Music } from "./music"
-import { Search } from "@/components/scene/Search"
+import { Music } from "./music"
+import { Audio } from "@/components/types/audio"
+import { Search } from "@/components/Search"
 import { Sidebar } from "@/components/Sidebar"
 import { useGlobalShortcut } from "@/hooks/tauri/shortcuts"
